requirements.txt
# python-telegram-bot==20.7
python-telegram-bot==22.0
requests==2.28.1
apscheduler==3.10.4
pytz==2023.3